requirements.txt
pillow
numpy
opencv-contrib-python
argparse
autopep8
flake8
pytest
tensorflow
streamlit
